+++ encode.py
@@ -13,7 +13,7 @@ def define_model(name, embed_mat, seq_len, funcs):
     vocab_num, embed_len = embed_mat.shape
     embed = Embedding(input_dim=vocab_num, output_dim=embed_len,
                       weights=[embed_mat], input_length=seq_len, trainable=True)
-    input = Input(shape=(seq_len,), dtype='int32')
+    input = Input(shape=(seq_len,))
     embed_input = embed(input)
     func = map_item(name, funcs)
     output = func(embed_input)
